hpccache: splitting and renaming the files
