+++ azurerm/internal/services/hpccache/validate/cache_namespace_path.go
@@ -1,0 +1,5 @@
+package validate
+
+func CacheNamespacePath(i interface{}, k string) (warnings []string, errs []error) {
+	return absolutePath(i, k)
+}
