+++ azurerm/internal/services/hpccache/validate/cache_nfs_export.go
@@ -1,0 +1,5 @@
+package validate
+
+func CacheNFSExport(i interface{}, k string) (warnings []string, errs []error) {
+	return absolutePath(i, k)
+}
