+++ azurerm/internal/services/hpccache/validate/cache_nfs_target_path.go
@@ -1,0 +1,5 @@
+package validate
+
+func CacheNFSTargetPath(i interface{}, k string) (warnings []string, errs []error) {
+	return relativePath(i, k)
+}
